% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-spatial.R
\name{as_spatial_cevodata}
\alias{as_spatial_cevodata}
\title{Spatial cevodata object}
\usage{
as_spatial_cevodata(object)
}
\arguments{
\item{object}{cevodata object}
}
\description{
Spatial cevodata object
}
