% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-spatial.R
\name{plot_space}
\alias{plot_space}
\alias{plot_empty_space}
\alias{layer_mutation_spread}
\title{Space plots}
\usage{
plot_space(object, mapping = aes(), ...)

plot_empty_space(object, fill = "gray90", ...)

layer_mutation_spread(object, snvs, ...)
}
\arguments{
\item{object}{cevodata object}

\item{mapping}{dmapping for geom_tile}

\item{...}{other parameters passed to geom_tile()}

\item{fill}{fill}

\item{snvs}{currently not used}
}
\description{
Space plots
}
