% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata-spatial.R
\name{validate_spatial_cevodata}
\alias{validate_spatial_cevodata}
\title{Validate spatial cevodata}
\usage{
validate_spatial_cevodata(object)
}
\arguments{
\item{object}{cevodata object}
}
\description{
Validate spatial cevodata
}
