% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevotools-package.R
\docType{package}
\name{cevotools-package}
\alias{cevotools}
\alias{cevotools-package}
\title{cevotools: Tools for Cancer Evolution Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Paweł Kuś \email{kpawel2210@gmail.com} (\href{https://orcid.org/0000-0002-4367-9821}{ORCID})

}
\keyword{internal}
